Generalized trial fragments
