+++ CrowderApp/app/src/main/java/com/example/crowderapp/views/trialfragments/CountTrialFragment.java
@@ -17,10 +17,8 @@ import com.example.crowderapp.controllers.ExperimentHandler;
 import com.example.crowderapp.controllers.callbackInterfaces.unPublishExperimentCallBack;
 import com.example.crowderapp.models.Experiment;
 
-public class CountTrialFragment extends Fragment {
+public class CountTrialFragment extends TrialFragment {
 
-    Experiment experiment;
-    private ExperimentHandler handler = ExperimentHandler.getInstance();
 
     public CountTrialFragment() {
 
@@ -29,39 +27,6 @@ public class CountTrialFragment extends Fragment {
     public static CountTrialFragment newInstance() {
         CountTrialFragment fragment = new CountTrialFragment();
         return fragment;
-    }
-
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-        setHasOptionsMenu(true);
-        super.onCreate(savedInstanceState);
-
-    }
-
-    @Override
-    public void onPrepareOptionsMenu(Menu menu) {
-        super.onPrepareOptionsMenu(menu);
-        // You can hide the state of the menu item here if you call getActivity().supportInvalidateOptionsMenu(); somewhere in your code
-        MenuItem menuItem = menu.findItem(R.id.more_item);
-        menuItem.setVisible(true);
-    }
-
-    @Override
-    public boolean onOptionsItemSelected(@NonNull MenuItem item) {
-        switch (item.getItemId()) {
-            case R.id.location_item:
-                Log.e("yo", "yo");
-                break;
-            case R.id.unpublish_item:
-                handler.unPublishExperiment(experiment.getExperimentID(), new unPublishExperimentCallBack() {
-                    @Override
-                    public void callBackResult() {
-                        getFragmentManager().popBackStack();
-                    }
-                });
-
-        }
-        return super.onOptionsItemSelected(item);
     }
 
     @Override
